Fixing state errors in fixed arena classroom

+++ v2/cloud_functions/cloud_fn_utilities/periodic_maintenance/hourly_maintenance.py
@@ -1,6 +1,7 @@
 from cloud_fn_utilities.gcp.pubsub_manager import PubSubManager
-from cloud_fn_utilities.globals import PubSub
+from cloud_fn_utilities.globals import PubSub, FixedArenaClassStates
 from cloud_fn_utilities.state_managers.fixed_arena_states import FixedArenaStateManager
+import time
 
 from cloud_fn_utilities.cyber_arena_objects.fixed_arena_class import FixedArenaClass
 
@@ -23,10 +24,16 @@ class HourlyMaintenance:
     def delete_expired(self):
         expired_classes = self.fa_state_manager.get_expired()
         for fixed_arena_class in expired_classes:
-            build_id = fixed_arena_class.key.name
-            if self.debug:
-                FixedArenaClass(build_id=build_id, debug=self.debug).delete()
-            else:
-                self.pub_sub_mgr.msg(handler=PubSub.Handlers.CONTROL,
-                                     cyber_arena_object=PubSub.CyberArenaObjects.FIXED_ARENA_CLASS,
-                                     action=PubSub.Actions.DELETE)
+            fac_state = fixed_arena_class.get('state', None)
+            if not fac_state:
+                FixedArenaClass(build_id=fixed_arena_class.key.name).mark_broken()
+                continue
+
+            if fac_state not in [FixedArenaClassStates.BROKEN.value, FixedArenaClassStates.DELETED.value]:
+                build_id = fixed_arena_class.key.name
+                if self.debug:
+                    FixedArenaClass(build_id=build_id, debug=self.debug).delete()
+                else:
+                    self.pub_sub_mgr.msg(handler=PubSub.Handlers.CONTROL,
+                                         cyber_arena_object=str(PubSub.CyberArenaObjects.FIXED_ARENA_CLASS.value),
+                                         build_id=build_id, action=str(PubSub.Actions.DELETE.value))
